[lldb][AMDGPU] Query architecture from ROCdbgapi agents
GPU core debugging initialized ROCdbgapi register handling with a hard-coded gfx942 architecture. On cores captured from newer AMDGPU targets, that mismatch made ROCdbgapi reject wave register queries with compatibility errors, so LLDB could not read the GPU PC and backtraces showed invalid frames.

This changes core-file initialization to query the architecture from ROCdbgapi's attached agents instead of hard-coding a device id. The helper is shared with live debugging so both paths use the same architecture discovery flow. GPU core threads also prefer the per-wave architecture when available, falling back to the process architecture only if that query fails.

The shared helper accepts the ROCdbgapi deallocator callback from the caller and uses that to release the agent list, instead of assuming free().

Test plan:

- git diff --check

- ninja tools/lldb/source/Plugins/Process/amdgpu-core/CMakeFiles/lldbPluginProcessAmdGpuCore.dir/ProcessAmdGpuCore.cpp.o tools/lldb/tools/lldb-server/Plugins/AMDGPU/CMakeFiles/lldbServerPluginAMDGPU.dir/LLDBServerPluginAMDGPU.cpp.o

- Attempted full GPU AMD test command from ~/clayborg/llvm-project/build/Debug; build failed first in unrelated dirty ProcessElfCore.cpp changes in the original worktree before tests ran.
