+++ lldb/include/lldb/Utility/AmdGpuAddressSpaces.h
@@ -9,14 +9,7 @@
 #ifndef LLDB_UTILITY_AMDGPUADDRESSSPACES_H
 #define LLDB_UTILITY_AMDGPUADDRESSSPACES_H
 
-#include "llvm/ADT/StringSwitch.h"
-#include "llvm/Support/Error.h"
-#include <amd-dbgapi/amd-dbgapi.h>
-#include <cassert>
 #include <cstdint>
-#include <cstdlib>
-#include <functional>
-#include <optional>
 
 namespace lldb_private {
 
@@ -56,107 +49,6 @@ enum class DW_ASPACE_AMDGPU : uint64_t {
   /// focused thread of execution
   private_wave = 6,
 };
-
-#define AMD_DBGAPI_ENUM_TO_CSTR(e)                                             \
-  case e:                                                                      \
-    return #e
-
-/// Convert an AMD debug API status code to a human-readable string.
-inline const char *AmdDbgApiStatusToString(amd_dbgapi_status_t status) {
-  switch (status) {
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_SUCCESS);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_FATAL);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_NOT_IMPLEMENTED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_NOT_AVAILABLE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_NOT_SUPPORTED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT);
-    AMD_DBGAPI_ENUM_TO_CSTR(
-        AMD_DBGAPI_STATUS_ERROR_INVALID_ARGUMENT_COMPATIBILITY);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_ALREADY_INITIALIZED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_NOT_INITIALIZED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_RESTRICTION);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_ALREADY_ATTACHED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_ARCHITECTURE_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_ILLEGAL_INSTRUCTION);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_CODE_OBJECT_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_ELF_AMDGPU_MACHINE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_PROCESS_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_PROCESS_EXITED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_AGENT_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_QUEUE_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_DISPATCH_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_WAVE_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_WAVE_NOT_STOPPED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_WAVE_STOPPED);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_WAVE_OUTSTANDING_STOP);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_WAVE_NOT_RESUMABLE);
-    AMD_DBGAPI_ENUM_TO_CSTR(
-        AMD_DBGAPI_STATUS_ERROR_INVALID_DISPLACED_STEPPING_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(
-        AMD_DBGAPI_STATUS_ERROR_DISPLACED_STEPPING_BUFFER_NOT_AVAILABLE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_DISPLACED_STEPPING_ACTIVE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_RESUME_DISPLACED_STEPPING);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_WATCHPOINT_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_NO_WATCHPOINT_AVAILABLE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_REGISTER_CLASS_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_REGISTER_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_LANE_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_CLASS_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_SPACE_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_MEMORY_ACCESS);
-    AMD_DBGAPI_ENUM_TO_CSTR(
-        AMD_DBGAPI_STATUS_ERROR_INVALID_ADDRESS_SPACE_CONVERSION);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_EVENT_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_BREAKPOINT_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_CLIENT_CALLBACK);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_CLIENT_PROCESS_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_SYMBOL_NOT_FOUND);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_REGISTER_NOT_AVAILABLE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INVALID_WORKGROUP_ID);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_INCOMPATIBLE_PROCESS_STATE);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_PROCESS_FROZEN);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_PROCESS_ALREADY_FROZEN);
-    AMD_DBGAPI_ENUM_TO_CSTR(AMD_DBGAPI_STATUS_ERROR_PROCESS_NOT_FROZEN);
-  }
-  assert(false && "unhandled amd_dbgapi_status_t value");
-}
-
-#undef AMD_DBGAPI_ENUM_TO_CSTR
-
-/// Run a command from the amd-dbgapi library and return an llvm::Error if not
-/// successful.
-///
-/// \param func A callable that returns an amd_dbgapi_status_t.
-/// \return An llvm::Error indicating success or failure with a descriptive
-///         message.
-inline llvm::Error
-RunAmdDbgApiCommand(std::function<amd_dbgapi_status_t()> func) {
-  amd_dbgapi_status_t status = func();
-  if (status != AMD_DBGAPI_STATUS_SUCCESS)
-    return llvm::createStringError(llvm::inconvertibleErrorCode(),
-                                   "AMD_DBGAPI_STATUS_ERROR: %s",
-                                   AmdDbgApiStatusToString(status));
-  return llvm::Error::success();
-}
-
-/// Get the AMD debug API log level from the AMD_DBGAPI_LOG_LEVEL environment
-/// variable. Returns the log level if set, or std::nullopt if not set.
-/// Valid values: "none", "fatal", "warning", "info", "verbose", "trace".
-inline std::optional<amd_dbgapi_log_level_t> GetAmdDbgApiLogLevelFromEnv() {
-  const char *env = std::getenv("AMD_DBGAPI_LOG_LEVEL");
-  if (!env)
-    return std::nullopt;
-  llvm::StringRef val(env);
-  return llvm::StringSwitch<std::optional<amd_dbgapi_log_level_t>>(val.lower())
-      .Case("none", AMD_DBGAPI_LOG_LEVEL_NONE)
-      .Case("fatal", AMD_DBGAPI_LOG_LEVEL_FATAL_ERROR)
-      .Case("warning", AMD_DBGAPI_LOG_LEVEL_WARNING)
-      .Case("info", AMD_DBGAPI_LOG_LEVEL_INFO)
-      .Case("verbose", AMD_DBGAPI_LOG_LEVEL_VERBOSE)
-      .Case("trace", AMD_DBGAPI_LOG_LEVEL_TRACE)
-      .Default(std::nullopt);
-}
 
 } // namespace lldb_private
